Change some files around
