+++ App.js
@@ -1,82 +1,49 @@
-/* src/App.js */
-import { withAuthenticator } from '@aws-amplify/ui-react'
-import React, { useEffect, useState } from 'react'
-import Amplify, { API, graphqlOperation } from 'aws-amplify'
-import { createTodo } from './graphql/mutations'
-import { listTodos } from './graphql/queries'
+/*--- 
+  SUMMARY:
+  This is the main file that the program will run 
 
-import awsExports from "./aws-exports";
-Amplify.configure(awsExports);
+---*/
+import React, { useState } from 'react';
+import * as Font from 'expo-font';
+import Home from './screens/home';
+import {AppLoading} from 'expo';
+import { AppNavigator } from './routes/drawer';
 
-const initialState = { name: '', description: '' }
 
-const App = () => {
-  const [formState, setFormState] = useState(initialState)
-  const [todos, setTodos] = useState([])
+const getFonts = () => Font.loadAsync({
+  'nunito-regular': require('./assets/fonts/Nunito-Regular.ttf'),
+  'nunito-bold': require('./assets/fonts/Nunito-Bold.ttf')
+});
 
-  useEffect(() => {
-    fetchTodos()
-  }, [])
+export var accounts = []
 
-  function setInput(key, value) {
-    setFormState({ ...formState, [key]: value })
-  }
-
-  async function fetchTodos() {
-    try {
-      const todoData = await API.graphql(graphqlOperation(listTodos))
-      const todos = todoData.data.listTodos.items
-      setTodos(todos)
-    } catch (err) { console.log('error fetching todos') }
-  }
-
-  async function addTodo() {
-    try {
-      if (!formState.name || !formState.description) return
-      const todo = { ...formState }
-      setTodos([...todos, todo])
-      setFormState(initialState)
-      await API.graphql(graphqlOperation(createTodo, {input: todo}))
-    } catch (err) {
-      console.log('error creating todo:', err)
+export class Account {
+    constructor(email, username, password) {
+        this.email = email
+        this.username = username
+        this.password = password
+        accounts.push(this)
     }
+}
+
+/*--- 
+This comment is a test comment to put while in testFile branch
+---*/
+
+export default function App(){
+  const [fontsLoaded, setFontsLoaded] = useState(false)
+  new Account("example@email.com", "user1", "123")
+
+  if(fontsLoaded){
+      return (
+          <AppNavigator />
+      );
+  } else {
+      return (
+          <AppLoading
+              startAsync={getFonts}
+              onFinish={() => setFontsLoaded(true)}
+          />
+      )
   }
-
-  return (
-    <div style={styles.container}>
-      <h2>Amplify Todos</h2>
-      <input
-        onChange={event => setInput('name', event.target.value)}
-        style={styles.input}
-        value={formState.name}
-        placeholder="Name"
-      />
-      <input
-        onChange={event => setInput('description', event.target.value)}
-        style={styles.input}
-        value={formState.description}
-        placeholder="Description"
-      />
-      <button style={styles.button} onClick={addTodo}>Create Todo</button>
-      {
-        todos.map((todo, index) => (
-          <div key={todo.id ? todo.id : index} style={styles.todo}>
-            <p style={styles.todoName}>{todo.name}</p>
-            <p style={styles.todoDescription}>{todo.description}</p>
-          </div>
-        ))
-      }
-    </div>
-  )
 }
-
-const styles = {
-  container: { width: 400, margin: '0 auto', display: 'flex', flexDirection: 'column', justifyContent: 'center', padding: 20 },
-  todo: {  marginBottom: 15 },
-  input: { border: 'none', backgroundColor: '#ddd', marginBottom: 10, padding: 8, fontSize: 18 },
-  todoName: { fontSize: 20, fontWeight: 'bold' },
-  todoDescription: { marginBottom: 0 },
-  button: { backgroundColor: 'black', color: 'white', outline: 'none', fontSize: 18, padding: '12px 0px' }
-}
-
-export default withAuthenticator(App)
